use cron to publish data
