+++ pubspec.yaml
@@ -34,6 +34,7 @@ dependencies:
   printing: ^3.7.2
   clock: ^1.0.1 # useful for testing
   permission_handler: ^5.0.1+1
+  cron: ^0.2.4
 
   # data-related
   provider: ^3.0.0
